+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Revision history for bytesmith
 
-## 0.3.2.0 -- 2019-??-??
+## 0.3.3.0 -- 2020-??-??
+* Add `hexWord8`, `hexWord16`, and `hexFixedWord8`.
+
+## 0.3.2.0 -- 2019-12-27
 
 * Add `parseBytesEither` and `parseBytesMaybe`.
 * Add common idioms from other parser libaries. This includes: `satisfy`,
   `satisfyWith`, `scan`, `peek`, and `peek'`.
-* Add `hexWord8`, `hexWord16`, and `hexFixedWord8`.
 
 ## 0.3.1.0 -- 2019-11-12